3-data_validation/requirements.txt
numpy==1.21.6
pandas==1.3.5
airflow-provider-great-expectations==0.1.5